+++ Modules/Models/Models/Client.swift
@@ -2,10 +2,10 @@ import Foundation
 
 public struct Client: Codable, Entity, Equatable {
     
-    public var id: Int
+    public var id: Int64
     public var name: String
     
-    public var workspaceId: Int
+    public var workspaceId: Int64
     
     enum CodingKeys: String, CodingKey {
         case id
