+++ Modules/Models/Models/Project.swift
@@ -2,15 +2,15 @@ import Foundation
 
 public struct Project: Codable, Entity, Equatable {
     
-    public var id: Int
+    public var id: Int64
     public var name: String
     public var isPrivate: Bool
     public var isActive: Bool
     public var color: String
     public var billable: Bool?
     
-    public var workspaceId: Int
-    public var clientId: Int?
+    public var workspaceId: Int64
+    public var clientId: Int64?
     
     enum CodingKeys: String, CodingKey {
