+++ Modules/Timer/TimerTests/TestUtils/TimeEntry+Extensions.swift
@@ -2,8 +2,13 @@ import Foundation
 import Models
 
 extension TimeEntry {
-    static func with(id: Int? = nil, description: String = "", start: Date? = nil, duration: TimeInterval? = nil, workspaceId: Int = 0) -> TimeEntry {
-        let timeEntryId = id ?? UUID().hashValue
+    static func with(
+        id: Int64? = nil,
+        description: String = "",
+        start: Date? = nil,
+        duration: TimeInterval? = nil,
+        workspaceId: Int64 = 0) -> TimeEntry {
+        let timeEntryId = id ?? Int64(UUID().hashValue)
         return TimeEntry(
             id: timeEntryId,
             description: description,
